Add Domain when Indicator is added via Object creation
When a domain or URL indicator is added using the standard indicator creation form, the domain is parsed out and added as a Domain in CRITs. This is also the case when clicking the '+' (Add Indicator) button next to an already existing Object. However, if the "Add Indicator" box of the Add Object form is selected during Object creation, the Domain is not parsed out and added. This behavior is inconsistent.