Removal of screenshots and storing screenshot ids.
You can now remove screenshots from a top-level object. When you mouse
over a thumbnail you'll get two icons:

- trash: remove the icon from the top-level object.
- ring/dot: store the screenshot id to a cookie for use in the upload
  screenshot form. you can store multiple ids. if the icon is a ring it
  means you can store it. if it stays permanently visible and is a dot,
  it means it's in your cookie ready to be used.

There is now a trash icon at the top of the Screenshots section. If it
is black it means there are not stored Screenshot IDs in your cookie. If
it is blue it means there are. If you click it, it will clear out all of
the stored Screenshot IDs in your cookie.