Merge tags on upload, style changes, etc.
When you upload a screenshot, if it existed before and you add tags in
your submission, we will now merge the tags with the existing tag list
so they aren't lost.

The screenshots now have a little box-shadow. If a screenshot was
assigned to the top-level object it will have a black box-shadow (also
helps with differentiating white in the image with the background of the
UI). If a screenshot was discovered through the bucket list tags, it
will show up with a blue box-shadow.

The Screenshots section now collapses properly.