Screenshot API and other fixes.
Added Screenshots to the API.

Had to make some changes to the core API to support images being added
to the zip files that it outputs so you can do ?file=1 on a screenshot
and get the image back.

You can read Screenshots and upload screenshots through the API. There
are two modes, "ids" and "screenshot". The former asumes you are
uploading a comma-separated list of ObjectIds of existing screenshots to
append to a top-level object. The latter assumes you are uploading an
image to the system.

Fixed some bugs with uploading screenshots with no description, no tags,
or no source information.