Screenshot listing and searching.
Hooked Screenshots up with a jTable listing page. It shows the
thumbnails in a column so you can easily see which image the row is
talking about. Tags are clickable allowing you to narrow down the
listing by tag. Clicking on the Details button gives you the full blown
image.

Due to leveraging jTable, Screenshots can now be marked as Favorites
(only from a listing page at this time).

You can also use global search against screenshots. It will default to
the description and tags fields of the Screenshot, but can be altered
using search operators like normal.

You can not subscribe to a Screenshot.