Major re-work of services.
The idea is to reduce the complexity involved in the services code by
pushing configuration and form creation out to services. This is very much
a WIP and still has more to do.

Big changes in this commit include:

The concept of a service_env is gone. Instead the manager is instaniated
just by importing crits.services. The concept of a database service is gone.

Importing, exporting and reseting service config is gone.

Many of the things that defined a service are removed. This includes type_
and required fields. The description is now explicitly defined by each
service and does not come from the docstring.

Make to_dict() in CritsDocumentFormatter actually return a dict and not a
SON object.