Initial attempt at tracking "sectors".
This commit adds the ability to track what sector(s) a top-level object
applies to. The list of available sectors is based off of the DHS
Critical Infrastructure Sectors list.

It is another field on the "Details" page for all top-level objects. It
uses the tagit jquery library (which we use for bucket lists and
campaign aliases). The list of available options is pre-populated to
choose from and drops down on field focus.

This also adds in sector count tracking similar to bucket lists as well
as searching.