Docstrings and bucket list autocompletion.
Added docstrings for some functions I missed. While I was here I quickly
added autocompletion for bucket list items. It will search the
"bucket_lists" collection for existing buckets as you type (minimum of
two characters) and find any that start with that string (case
insensitive). It will return that list as available autocomplete options
to the user in the UI.