Strip whitespace when creating Indicator from Object
Before Indicators are added, they are converted to lowercase and whitespace is stripped. If this is not first done in the 'create_indicator_from_object' function, an Object with uppercase characters or leading/trailing whitespace will cause the relationship to not be formed because the Object value and Indicator value will not match.