Fix long-standing config in template bug.
When viewing service analysis results the configuration was not showing up.
This is because in the move to use mongoengine the config is now an
AnalysisConfig object and the template was expecting a dictionary.

Fix this by implementing a to_dict filter that can be used in any template.
It will call to_dict() on whatever is passed into it. Make
service_analysis_widget template use this.