Sort Download Objects, and fix CybOX "raw" error.
The list of Objects you can choose to download when using the Download
feature is now alphabetized.

CybOX throws an error when you try to Download in STIX format and select
a binary format of "raw" under certain conditions (like null bytes). The
form will now force "base64" on the user if they select STIX as their
download format.