Start of Actor top-level object.
Lots of stuff going on but little to not functionality in the UI is
here.

Only thing you'll see in the UI is the new Items necessary to drive
Actor input. Those are visible in the Control Panel and there's a
management command to populate them (which is now hooked into
create_default_collections).

Generic views and handlers are written but not used at the moment. Forms
are started. Basically need to build up the UI now and then slowly add
in all of the UI and backend code for adding/editing/removing.