Fixing error with standards API upload and source addition.
The standards API needed an 'object_class' defined in meta that it could
use (even though we don't need it yet).

Also there was an issue with source addition when ingesting standards
documents. For some reason a change that we made to fix double-source
uploading magically resulted in no source uploading. I've added the
additions back to fix the problem but have note tested this when
ingesting from TAXII (TAXII server I test with is currently down).