Hooking up Actor Identifiers interfaces.
Ability to view a listing page and add new Actor Identifiers to the
system. These will in turn have the ability to be attributed to an Actor
top-level object (that piece is in the works).

If you remove an Actor Identifier from the system it will remove its
attribution from all Actors it's attributed to.