Give samples ability to have multiple filenames.
Works similar to campaign aliases. There is one main filename (which can
now be modified in-line) which is the filename your organization prefers
that sample go by. There's also a list of filenames which are other
filenames people have reported that binary going by.

If you upload a Sample that already exists in the database, the filename
will get appended to the list of filenames if it isn't already the
filename or in the list of existing filenames.

Searching "Filename" in the advanced search for Samples now does a $or
between both fields so you will search both simultaneously.

Closes #5.