Hooking up actor name, description, aliases, attribution.
Attributing Identifiers to Actors is now hooked up. Currently you need
to add an Identifier before you can go attribute it. Some day in the
future we may add it such that you can add new identifiers on the fly
while performing attribution.

You can now edit the name, aliases, and description of an Actor. Aliases
do not prevent you from adding a value that exists as a Name or Alias of
another Actor. This is due to the potential for naming conflicts across
origins. Having the ability to edit the name and aliases list allows you
to resolve that conflict how you see fit.

Due to the amount of url patterns we had, I ran into the situation where
we were passing more than 255 arguments to `patterns()` in
`crits/urls.py`. It's been a long time coming, but now each TLO has its
own urls.py file. This is a large looking change but changes nothing
functionally. It just makes it cleaner and easier to add urls and know
where to find them.

I believe this adds all of the basic functionality for an Actor with the
exception of pivoting off of an Identifier value to find TLOs which use
that value somewhere.