Hook up pivoting on identifier_id.
From the "Identifiers" section and from the Identifiers listing page,
you can now click on the "Name" of the Identifier and it will link you
to the Actors listing page with all of the Actors which have that
Identifier attributed to them.

Note: I am not 100% positive I hooked this up correctly, but it seems to
work. jtable searching is crazy :(