Hook up STIX support for Actors.
This will allow you to export and import an Actor in STIX format. The
only piece missing right now is including Identifiers. I am waiting to
hear back from some developers on where the best place to put those
would be.

Hooked up the downloading interface for Actors and added the ability to
Delete an actor from the Details page.