Fixing service_run logic.
If there was no form object created it would be None and fall into the
if POST condition and try to instantiate None with the request data
throwing an error.

Fixed some roundabout logic where if get_form had no form to give it
would return a redirect, but the javascript was removed for detecting
and using the redirect. Now instead of relying on (missing) JavaScript
we now call service_run with the appropriate information for you and
return the results of that in the response (which the JavaScript does
support).