Fix email API field upload.
Default non-required fields to sane values. This now allows users to use
the API to upload an email as fields. Any fields which are sent which are
not supported by the class will be put into unsupported_attrs.

Strip the username and API key out of the bundle before passing to
handle_email_fields() so that they are not put in unsupported_attrs.