Be consistent in getting services.
Instead of calling into the manager to get enabled services just use
the handler instead. This gets us one step closer to not needing to
restart the webserver after adding/removing services.

This is also consistent with how triage services are obtained.

Now that the manager is not being used to provide these functions they
are removed.