Fix pcap upload handler.
This is an identical commit to b6f073a930242cbb7ab216a49175fac2b67e9eed.

Make it so if related_id or related_md5 is set you must also
set related_type and vice versa. And if the related object can not be
found complain early.

This ensure that if relationships are specificed but can not be made we
will refuse to add the PCAP.

Also, fix a copy/paste error in a docstring.