Make class_from_id() behave better.
While playing with the API I discovered that class_from_id() would
raise an exception if the ID is not a valid object ID format (whatever
that is). I was testing and gave it an ID of 'foo', which raised a
ValidationError exception. To fix this I've made class_from_id catch that
exception and return None instead.

I also walked all the places where class_from_id() is used in core
and made sure they properly error out when class_from_id() returns None.