Allow for unicode characters in Audit value.
We allow for unicode characters in other places (such as Campaign name)
but this was attempting to cast it to a string and failing. I tested
with the other types to make sure audit entry values were being
generated properly (specifically the ones that use ObjectIds) and they
seem to all be working properly.

Closes #224.