Allow for returning a single STIX document.
Changes allows code to use the download object handler to return a
single STIX document instead of a zip of multiple STIX documents.
Default behavior is still a zip of multiple STIX documents.

Issue with this comes when there's an Event or multiple events. Since
each Event is its own unique STIX document, you can't have more than one
Event in the STIX doc. Also, this code doesn't analyze all of the
objects it needs to convert to find an Event to base the document off
of.