Start of Roles.
Adds a basic Role class with a ton of attributes around all of the
features for CRITs. Current support allows for adding a source, adding
all sources, and making all attributes True.

User class was updated to allow for us to get a total Access List for
that user, as well as lookup access by individual access attribute. The
roles are landing in a new "roles" attribute. Bonus that it allows me to
test this code without affecting the current ACL structure.

Default role added to create_default_collections.

Only TLO that has ACL attributes is Actor at the moment. I'm using it as
my POC. Once I feel everything is working "as intended" I'll extend
support to the other TLOs.

This should have 0 effect on the UI as nothing is leveraging this yet
and there's no interface to add/edit/disable Roles.