Another helper method.
Added a "can_do_one_of()" method which will allow you to pass in a list
of ACLs and determine if the user has access to at least one of them.
This will be useful for determining if we should populate a form in the
UI (like the Add Relationships form) where the user has access to use it
in some places but not others.

Added an "update" option to "get_access_list()" which will now returned
the cached_acl if it's populated instead of always updating the cache.
You can set "update" to True and it will update the cache before
returning. Should save some processing in a couple situations.