Make analysis results their own collection.
These code changes make analysis results go to their own collection. The
upgrade script is necessary for doing the migration but you can freely
spin up the UI and have things migrated on-access.

This change will allow services to utilize an entire document on their
own per-run, giving them almost 16MB of space for results instead of
having to share results with other services run for the same TLO.

This commit does not include the following things that will be done
soon:
    - An Analysis Results listing page.
    - Removing the "Yara Hits" page (still questionable).
    - Add an Analysis Results tab to global search which shows the TLO
      type, service name, service version, date of execution, number of
      results, and a link to the TLO.
        - Need to figure out how to limit these by source as well.
    - An entry in UPDATING letting people know what's going on.