Default API key for users.
New users will have an API key auto-generated for them. It will be set
as their Default key. This key will be used by CRITs for situations
where the user does not have the ability to enter an API key (such as
running services on triage).

Users have the ability to change which API key is their default one.
Users cannot revoke an API key if it is marked as their default.