Start building Roles interface.
Roles now have their own section in the Control Panel. You can view the
roles and change Active state in the jtable. The details link will open
up a page where you can review and modify all of the items in the Role.

On the Details page you get tabs for all of the different sections. You
cannot modify the name of a Role (currently). The description field will
eventually be editable as well as the source list. Next to the name is a
"Copy From" link which allows you to copy the Role and all of its
settings to a Role with a new name.