Append force_full=1 to pivot links.
If you have a service result which is "foo (bar)" the parenthesis are
taken as part of the regex, and the pivot is broken. The 'detectexact'
notion was killed off in d240572d01482b4f6dd0b92f45c1310abb334479, which
is what I think is the root cause of this.