Hook up sources interface.
Interface allows you to manipulate the sources that are associated with
a role.

It uses the multiselect interface so you can move sources between being
added/removed. When a source is added, it will add a new interface for
you to choose the different access items for that source. If you remove
a source its interface will go away.

"Add All" will add all of the sources but will not pre-configure access
for any of them. "Remove All" will remove them all completely (current
bug: the interface leaves one row behind even though it is in fact
removed).