Edit name, description, and role combination preview.
You can now edit the name and description of a Role. If you change the
name it will look for any users who are assigned to that Role and change
the name accordingly.

There's also a Role Combinator. This allows you to choose a list of
Roles and have them combined to show you what the resulting access would
look like. It uses the Role Details page to render the results so it
should be very simple to browse like any other Role.