Remove old role and administrator  access controls.
With this commit the entire system is freely open to all users. In order
to start clean and implement roles I wanted to first remove all
instances where we are using the old user role attribute and limiting
anything by the "Administrator" concept.

"user_sources()" is still a thing and content is limited by it but it
now determines the list based off of the sources assigned to the roles
you have.

Removed the old "Role" and "Sources" sections for Users in the Control
Panel and replaced it with the "Roles" interface.

At this point if you want to assign sources to people you have to do so
through roles.