Add source to UberAdmin role if it's created.
When adding a new source to the system it will automatically try to add
it to the UberAdmin role to keep it updated with the latest source
access.

Also removed the "UberAdmin" string and replaced it with a settings
variable ADMIN_ROLE if people are so inclined to change the name.

Fixed a bug where you couldn't add a new source because the form wasn't
rendering due to old code.