Fix some problems with domains API.
This fixes a problem where you add a domain via the API and choose to also
add an IP with it. The ip_add_update() handler requires that you pass an
ip_type that is not None. Fix this by taking an ip_type parameter in the
POST data and passing it along.

Another problem is that the handler would not error out early if there
is no source provided. This would cause a validation error later on.

Also, fix some whitespace problems.