Set Campaign confidence to 'low' if unexpected value is provided
It seems that this base function should enforce the three "low", "medium", or "high" options for Campaign Confidence. If the provided value is not one of those three, the code should either throw an error, or set the value to the default of "low".

I went with the latter option in this case. What do you think?