Fix login screen.
Alignment for the login screen was all sorts of messed up after the
change which removes the TOTP options if it's disabled. It was further
messed up in Safari.

This should align things in the center across browsers (tested Safari,
Chrome, FF all on OSX) and should look the same whether TOTP is enabled
or not.