Better debugging when debug mode is not turned on.
This change exposes the Exception name and the Value in the custom
500.html template. Created an "errors.py" file which contains the
custom_500 function to facilitate populating the template. This function
also formats the traceback and write it to the "crits.log" file as
DEBUG. This will allow people to disable DEBUG mode, but set logging to
DEBUG. Now the UI will be nice and clean, users will have a small bit of
information which they can relay to an admin who can find the
corresponding information in the log.