Make services use thread local storage.
When running multiple services on an object in a threaded model there would
occasionally be "odd" behavior. At best it would be incorrect results from
a service, and at worst it would be a deadlock. Here's an example:

Let's say we have YARA and peinfo set to run on triage. If I add a sample
a thread would be kicked off for each service. If YARA reads from the object
first then peinfo reads the second read would be sometimes result in a read
from some random part of the file. Occasionally this scenario would result
in a deadlock.

By making each object passed to a service be thread-local we side-step the
entire problem, at the cost of slightly more memory usage.