Make TLP selection work properly.
If you attempt to select a TLP level less restrictive than the
least-restrictive source instance it will not change the TLP level and
will show you a warning.

If you select an equal or more restrictive TLP level, you will get a
success.

I was going to limit the dropdown to only those acceptable values but
did not because:
- adding a new source instance will alter which options are available
- people can hand-craft the request regardless of what is in the list
- if someone adds a source instance to the same TLO and you haven't
  refreshed the page the options will still be available.