Modify query so Comments appear on Recent Activity page
Comments were not appearing in the Recent Activity (Aggregate Comments) page.
Back in May 2014, notifications were moved out of the Comments collection of the DB, which eliminated the need for the 'type' field. From that point forward, adding a new Comment would not include a 'type' field, but the "get_aggregate_comments" function still queried for data using `{'type': 'comment'}`.

Modifying the query fixes this problem.