Fix requirements and bootstrap.
I added a dupe for pyyaml in requirements so I removed one. Also,
bootstrap was blindly downloading and installing MongoDB every time.
Added a check for the existence of MongoDB and it will only install if
it's not found. This check doesn't do a version check (which is probably
better) but it's better than nothing.

This puts us more in-line for supporting installing CRITs via Homebrew.