Attempt to make bootstrap a complete setup script.
I was able to go from zero-to-hero with this on OSX. I will note I had
all of the dependencies installed already so there might be tweaks
necessary to the list for brew.

Added a script to generate a secret key used by bootstrap when creating
the default database file. Also created a simpler MongoDB start script.
Regardless of the system you are on it will work. It may not be 100%
performant since we're not paying any attention to NUMA, but if you are
just looking to get a system up and running, this will work.