Standardizing handler argument names.
The patch API needs to be able to determine the type and id
automatically from the request URL and not rely on the user providing
that information in the request data. This requires us to know where in
the data dictionary to stuff the id once we've derived it. To do this
handlers need to be standardized so that common argument names are
always the same. This currently includes:

- id_
- user

The user is important for patch as we don't want someone to be able to
pass in a different user. We need to be able to overwrite any user they
might try to pass in with the authenticated user making the request.