Fix a couple bugs and change value generation.
The value of the object being generated in the object_mapper was
inconsistent due to the different types of objects a CybOX object can
return when acquiring the "value". Given that there are many which are
StructuredText, or other types of objects, I'm adding a simple function
which tries to get the ".values" property of the attribute in question
and if that fails just return a list with the ".value" attribute. This
should cover the inconsistencies but also be flexible enough to
accommodate any changes to CybOX in the future where the attributes are
changed to allow for ".values" to work.