Fixing b64decode and encode/decode stuffs.
There were issues with us attempting to b64decode stuff that was already
b64decoded by the STIX/CybOX libraries and it caused weirdness. After
removing those I also noticed that the Artifact and File objects didn't
have descriptions so they were failing to generate. I set the
descriptions to "None" so at least the user can modify it in the UI
post-upload.

Also fixed an issue with RawData where it would fail to import because
we do a ".encode('utf-8')" in handlers but it was already a string going
into that function if doing a STIX import. Sadly, I had to add
".decode('utf-8')" which is an extra set of weirdness but necessary
given that RawData can come into that function in uncontrolled
encodings.