Initial stab at adding Location tracking for TLOs.
This adds the ability to track location information for a TLO,
specifically where it originated from and/or where it is/was destined
to. The originating and destined locations can have many meanings, we do
not require the meaning. There is a description field where a meaning
can be added so it's clear what the location pertains to.

The location is based on Country. There is a management command
"create_locations" to add these to the system.

It also allows for the user to add in the latitude and longitude of the
location if they wish to be more specific.