If we cannot email the user, don't die horribly.
If the user is about to break the treshold of invalid login attempts,
the system will attempt to send them an email letting them know their
account is being locked out.

If the system can't send that email, an exception is thrown, the user's
account is not set inactive, and their invalid login attempts is not
incremented properly.

This catches the exception and logs it so an admin can debug further.
The user's account is now properly updated.