Add support for backdoors to sample upload.
When uploading a sample the form now has a "Backdoor" entry, which is
populated with all the available backdoors the user has source access
to.

The listing of available backdoors is done with a new handler called
"get_backdoor_names" which returns a list of tuples containing backdoor
name and backdoor version. The form init converts these to be a
unique-ish value for submission (name|||version). Upon submission of the
form the unique value is broken back out into the individual parts and
acted on accordingly.

If you select a specific instance of a backdoor the sample will be
related to both the family and the specific version. If you select only a
family then the sample will be related only to the family.

This does require passing the backdoor name and backdoor version along
the handle_upload_file path, including all the other functions it calls.