Fix broken default dashboards.
Now that backdoor and exploit are no longer attributes of samples the
"Top_Backdoors" default dashboard was removed. However, the code in
getRecordsForDefaultDashboardTable() was removed in an earlier commit by
me but the code was missing an else clause that would produce an "empty"
result set for the default "Top_Backdoors" dashboard.

Fix this by adding an else clause so that if default dashboards are
removed in the future they won't break things. I intend to re-add the
"Top_Backdoors" default dashboard in a future commit so people do not
have to recreate the default dashboards to remove it.